// origin/src/youtube_music/types/ArtistResults_1.ts
export interface ArtistResults_1 {
    responseContext: ResponseContext
    contents: Contents
    header: Header2
    trackingParams: string
    frameworkUpdates: FrameworkUpdates
}

export interface ResponseContext {
    serviceTrackingParams: ServiceTrackingParam[]
}

export interface ServiceTrackingParam {
    service: string
    params: Param[]
}

export interface Param {
    key: string
    value: string
}

export interface Contents {
    singleColumnBrowseResultsRenderer: SingleColumnBrowseResultsRenderer
}

export interface SingleColumnBrowseResultsRenderer {
    tabs: Tab[]
}
// ...
export interface Header2 {
    musicHeaderRenderer: MusicHeaderRenderer
}

export interface MusicHeaderRenderer {
    title: Title5
    trackingParams: string
}

export interface Title5 {
    runs: Run12[]
}

export interface Run12 {
    text: string
}

export interface FrameworkUpdates {
    entityBatchUpdate: EntityBatchUpdate
}

export interface EntityBatchUpdate {
    mutations: Mutation[]
    timestamp: Timestamp
}